pyproject.toml
[project]
name = "database-reviews"
version = "0.1.0"
description = ""
authors = [
    {name = "Liza Katz",email = "lizka.k@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "numpy (>=2.4.0,<3.0.0)",
    "elasticsearch (>=9.2.1,<10.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain (>=1.2.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "cohere (>=5.20.1,<6.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pinecone (>=8.0.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode=false
[tool.poetry.group.dev.dependencies]
nbstripout = "^0.8.2"
pre-commit = "^4.5.1"

